requirement.txt
MetaTrader5
pandas
numpy
scikit-learn
tensorflow
openpyxl
matplotlib
seaborn
ta
joblib
optuna
requests
streamlit
python-telegram-bot